requirements.txt
Flask==2.3.2
python-docx==0.8.11
spacy==3.5.0
en-core-web-sm==3.5.0
googlesearch-python==1.0.1
gunicorn==20.1.0